market_analyzer/requirements.txt
Django==4.0
djangorestframework==3.13.1
django-markdown==0.8.4
django-filter==2.1
django-request==1.5.1
python-dateutil==2.8.2
